move file to lib/util
